feat: Add format-specific annotations to override secret file names (#572)
* feat: Add format-specific annotations to override secret file names

* docs: Add annotations to volume page

* test: Add custom secret name tests

* feat: Add check against path traversal

* refactor: Flatten struct, use moved value

* refactor: Flatten compat struct, use moved value

* test: Add unit test for selector deserialization

* refactor: Adjust path traversal checks

Path::canonicalize will return an error if the path does not exist.
The path we are checking obviously doesn't exist yet, because we want
to prevent path traversals and the file at that path will only exist
after we are done with the check. So using canonicalize does not work
in this use-case.

* test: Add custom CA name

* chore: Use quoted paths in error messages

* refactor: Ensure all components of a path a normal

* chore: Add changelog entry