test: Fix CRD conversion failures
The current testing script immediately continues to run tests
after all required operators are installed. This can lead to CRD
conversion failures because all resources (this includes custom
ones) are applied by kuttl immediately after. All custom resources
(SecretClasses in this case) use v1alpha1 (to additionally test
the conversion) and as such cannot be converted to the stored
v1alpha2 version when the conversion webhook is not running yet.

To remedy this, a 10 second wait is introduced after the operator
installation. A far better solution is to wait for a Kubernetes
status condition, but that is more involved. I would like to stop
bolting on functionality to the test script and instead move these
features into our new (hopefully coming soon) tooling.